backend/requirements.txt
# Flask Core
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
gunicorn==21.2.0

# Database
psycopg2-binary==2.9.9

# Calculation Engine
pandas==2.1.3
Pint==0.23

# Auth
PyJWT==2.8.0

# Utilities
python-dotenv==1.0.0

# Testing
pytest==7.4.3